Update dashboard layout and builder page
