+++ app/dashboard/layout.tsx
@@ -1,9 +1,11 @@
-'use client'
+"use client"
+
 import Sidebar from '@/components/Sidebar'
 import { useUser } from '@/hooks/useUser'
 
 export default function DashboardLayout({ children }: { children: React.ReactNode }) {
   const { user } = useUser()
+
   return (
     <div className="flex min-h-screen">
       <Sidebar />
@@ -11,7 +13,9 @@ export default function DashboardLayout({ children }: { children: React.ReactNod
         <header className="p-4 border-b text-right text-sm text-gray-600">
           {user?.email}
         </header>
-        <main className="flex-1 p-8">{children}</main>
+        <main className="flex-1 p-8">
+          {children}
+        </main>
       </div>
     </div>
   )
